+++ src/commands/serve-react.js
@@ -2,16 +2,17 @@ import path from 'path'
 
 import glob from 'glob'
 
+import {UserError} from '../errors'
 import serveReact from '../serveReact'
 
-export default function(args) {
+export default function(args, cb) {
   if (args._.length === 1) {
     console.log(`usage: nwb serve-react <entry module>`)
   }
+
   let entry = args._[1]
   if (glob.sync(entry).length === 0) {
-    console.error(`entry module not found: ${path.join(process.cwd(), entry)}`)
-    process.exit(1)
+    return cb(new UserError(`entry module not found: ${path.join(process.cwd(), entry)}`))
   }
 
   console.log('nwb: serve-react')