+++ src/constants.js
@@ -4,4 +4,4 @@ export const REACT_APP = 'react-app'
 export const REACT_COMPONENT = 'react-component'
 export const WEB_MODULE = 'web-module'
 
-export const MODULE_TYPES = [REACT_APP, REACT_COMPONENT, WEB_MODULE]
+export const PROJECT_TYPES = [REACT_APP, REACT_COMPONENT, WEB_MODULE]