+++ src/getUserConfig.js
@@ -2,8 +2,9 @@ import path from 'path'
 
 import glob from 'glob'
 
-import {MODULE_TYPES} from './constants'
+import {PROJECT_TYPES} from './constants'
 import debug from './debug'
+import {UserError} from './errors'
 
 export default function getUserConfig(args = {}) {
   // Try to load default user config, or user a config file path we were given
@@ -11,8 +12,7 @@ export default function getUserConfig(args = {}) {
   let userConfigPath = args.absConfig || path.join(process.cwd(), args.config || 'nwb.config.js')
 
   if (glob.sync(userConfigPath).length === 0) {
-    console.error(`nwb: couldn't find a config file at ${userConfigPath}`)
-    process.exit(1)
+    throw new UserError(`nwb: couldn't find a config file at ${userConfigPath}`)
   }
 
   try {
@@ -20,19 +20,18 @@ export default function getUserConfig(args = {}) {
     debug('imported config module from %s', userConfigPath)
   }
   catch (e) {
-    console.error(`nwb: couldn't import the config file at ${userConfigPath}`)
-    console.error(e.stack)
-    process.exit(1)
+    throw new UserError(`nwb: couldn't import the config file at ${userConfigPath}`)
   }
 
   if (typeof userConfig == 'function') {
     userConfig = userConfig()
   }
 
-  if (MODULE_TYPES.indexOf(userConfig.type) === -1) {
-    console.error(`nwb: invalid module type configured in ${userConfigPath}: ${userConfig.type}`)
-    console.error(`nwb: 'type' config must be one of: ${MODULE_TYPES.join(', ')}`)
-    process.exit(1)
+  if (PROJECT_TYPES.indexOf(userConfig.type) === -1) {
+    throw new UserError(
+      `nwb: invalid project type configured in ${userConfigPath}: ${userConfig.type}`,
+      `nwb: 'type' config must be one of: ${PROJECT_TYPES.join(', ')}`
+    )
   }
 
   // If the user provided some Babel config, automatically apply it to