+++ tests/fixtures/invalid-config.js
@@ -1,0 +1,2 @@
+// A module which will fail when we try to require it
+'fail'.badMethod()